// server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  HOUSE_USER: 'HOUSE_USER',
  HOUSE_LANDLORD: 'HOUSE_LANDLORD',
  HOUSE_ADMIN: 'HOUSE_ADMIN',
  HOUSE_IT_ADMIN: 'HOUSE_IT_ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Language = {
  EN: 'EN',
  IT: 'IT'
} as const

export type Language = (typeof Language)[keyof typeof Language]